feat: wire up db storage for credit tracker
- Move creditrepov2 code into creditrepo package
- Add new events package for event handling
- Update gRPC definitions and generated code
- Add database test helpers
- Update dependencies and configurations

+++ pkg/creditservice/metrics.go
@@ -26,7 +26,7 @@ var (
 )
 
 // getAmountBucket returns a string label for the amount bucket
-func getAmountBucket(amount int64) string {
+func getAmountBucket(amount uint32) string {
 	switch {
 	case amount == 1:
 		return "1"
